+++ rummage/lib/gui/settings/__init__.py
@@ -22,10 +22,9 @@ from __future__ import unicode_literals
 import codecs
 import json
 import os
-import traceback
 import copy
 from datetime import datetime
-from . import portalocker
+from filelock import FileLock
 from ..app import custom_app
 from ..app.custom_app import debug, debug_struct, error
 from ..generic_dialogs import errormsg
@@ -72,13 +71,12 @@ class Settings(object):
         cls.ERR_SAVE_CACHE_FAILED = _("Failed to save cache file!")
 
     @classmethod
-    def load_settings(cls, debug_mode):
+    def load_settings(cls):
         """Load the settings."""
 
-        cls.debug = debug_mode
+        # cls.debug = debug_mode
         cls.localize()
-        cls.settings_file, cls.cache_file, log = cls.get_settings_files()
-        custom_app.init_app_log(log)
+        cls.setup_setting_files()
         cls.settings = {"__format__": SETTINGS_FMT}
         cls.cache = {"__format__": CACHE_FMT}
         cls.settings_time = None
@@ -87,8 +85,8 @@ class Settings(object):
         if cls.settings_file is not None:
             cls.open_settings()
             cls.open_cache()
-        if cls.debug:
-            custom_app.set_debug_mode(True)
+        # if cls.debug:
+        #     custom_app.set_debug_mode(True)
         localization.setup('rummage', os.path.join(cls.config_folder, "locale"), cls.get_language())
         cls.localize()
         debug_struct(cls.settings)
@@ -101,6 +99,12 @@ class Settings(object):
 
         cls.reload_settings()
         return copy.deepcopy(cls.settings)
+
+    @classmethod
+    def get_log_file(cls):
+        """Get location of log file."""
+
+        return cls.log
 
     @classmethod
     def is_regex_available(cls):
@@ -253,7 +257,7 @@ class Settings(object):
         return changed
 
     @classmethod
-    def get_settings_files(cls):
+    def setup_setting_files(cls):
         """Get settings, cache, log, and fifo location."""
 
         platform = util.platform()
@@ -265,9 +269,9 @@ class Settings(object):
             plugin_folder = os.path.join(folder, 'plugins')
             if not os.path.exists(plugin_folder):
                 os.mkdir(plugin_folder)
-            settings = os.path.join(folder, SETTINGS_FILE)
-            cache = os.path.join(folder, CACHE_FILE)
-            log = os.path.join(folder, LOG_FILE)
+            cls.settings_file = os.path.join(folder, SETTINGS_FILE)
+            cls.cache_file = os.path.join(folder, CACHE_FILE)
+            cls.log = os.path.join(folder, LOG_FILE)
             cls.fifo = os.path.join(folder, '\\\\.\\pipe\\rummage')
             cls.config_folder = folder
         elif platform == "osx":
@@ -281,9 +285,9 @@ class Settings(object):
             plugin_folder = os.path.join(folder, 'plugins')
             if not os.path.exists(plugin_folder):
                 os.mkdir(plugin_folder)
-            settings = os.path.join(folder, SETTINGS_FILE)
-            cache = os.path.join(folder, CACHE_FILE)
-            log = os.path.join(folder, LOG_FILE)
+            cls.settings_file = os.path.join(folder, SETTINGS_FILE)
+            cls.cache_file = os.path.join(folder, CACHE_FILE)
+            cls.log = os.path.join(folder, LOG_FILE)
             cls.fifo = os.path.join(folder, FIFO)
             cls.config_folder = folder
         elif platform == "linux":
@@ -293,41 +297,34 @@ class Settings(object):
             plugin_folder = os.path.join(folder, 'plugins')
             if not os.path.exists(plugin_folder):
                 os.mkdir(plugin_folder)
-            settings = os.path.join(folder, SETTINGS_FILE)
-            cache = os.path.join(folder, CACHE_FILE)
-            log = os.path.join(folder, LOG_FILE)
+            cls.settings_file = os.path.join(folder, SETTINGS_FILE)
+            cls.cache_file = os.path.join(folder, CACHE_FILE)
+            cls.log = os.path.join(folder, LOG_FILE)
             cls.fifo = os.path.join(folder, FIFO)
             cls.config_folder = folder
 
-        if not os.path.exists(settings):
-            cls.new_settings(settings)
-        if not os.path.exists(cache):
-            cls.new_cache(cache)
+        cls.settings_lock = FileLock(cls.settings_file + '.lock')
+        cls.cache_lock = FileLock(cls.cache_file + '.lock')
 
-        return settings, cache, log
+        if not os.path.exists(cls.settings_file):
+            cls.new_settings(cls.settings_file)
+        if not os.path.exists(cls.cache_file):
+            cls.new_cache(cls.cache_file)
 
     @classmethod
     def open_settings(cls):
         """Open settings file."""
 
+        success = False
         try:
-            locked = False
-            with codecs.open(cls.settings_file, "r", encoding="utf-8") as f:
-                assert portalocker.lock(f, portalocker.LOCK_SH), "Could not lock settings."
-                locked = True
-                cls.settings = json.loads(f.read())
-                assert portalocker.unlock(f), "Could not unlock settings."
-                locked = False
-                cls.update_settings()
+            with cls.settings_lock.acquire(2):
+                with codecs.open(cls.settings_file, "r", encoding="utf-8") as f:
+                    cls.settings = json.loads(f.read())
+            success = True
         except Exception:
-            e = traceback.format_exc()
-            try:
-                error(e)
-                if locked:
-                    portalocker.unlock(f)
-                errormsg(cls.ERR_LOAD_SETTINGS_FAILED)
-            except Exception:
-                print(str(e))
+            errormsg(cls.ERR_LOAD_SETTINGS_FAILED)
+        if success:
+            cls.update_settings()
 
     @classmethod
     def new_settings(cls, settings):
@@ -336,19 +333,12 @@ class Settings(object):
         default_settings = {'__format__': SETTINGS_FMT}
 
         try:
-            locked = False
-            with codecs.open(settings, "w", encoding="utf-8") as f:
-                assert portalocker.lock(f, portalocker.LOCK_EX), "Could not lock file."
-                locked = True
-                f.write(json.dumps(default_settings, sort_keys=True, indent=4, separators=(',', ': ')))
-                assert portalocker.unlock(f), "Could not unlock file."
-                locked = False
+            with cls.settings_lock.acquire(2):
+                with codecs.open(settings, "w", encoding="utf-8") as f:
+                    f.write(json.dumps(default_settings, sort_keys=True, indent=4, separators=(',', ': ')))
         except Exception:
-            try:
-                if locked:
-                    portalocker.unlock(f)
-            except Exception:
-                pass
+            # It's okay that we couldn't do it this time.
+            pass
 
         return default_settings
 
@@ -395,47 +385,32 @@ class Settings(object):
 
     @classmethod
     def open_cache(cls):
-        """Open cache file."""
+        """Open settings file."""
 
+        success = False
         try:
-            locked = False
-            with codecs.open(cls.cache_file, "r", encoding="utf-8") as f:
-                assert portalocker.lock(f, portalocker.LOCK_SH), "Could not lock settings."
-                locked = True
-                cls.cache = json.loads(f.read())
-                assert portalocker.unlock(f), "Could not unlock settings."
-                locked = False
-                cls.update_cache()
+            with cls.cache_lock.acquire(2):
+                with codecs.open(cls.cache_file, "r", encoding="utf-8") as f:
+                    cls.cache = json.loads(f.read())
+            success = True
         except Exception:
-            e = traceback.format_exc()
-            try:
-                error(e)
-                if locked:
-                    portalocker.unlock(f)
-                errormsg(cls.ERR_LOAD_CACHE_FAILED)
-            except Exception:
-                print(str(e))
+            errormsg(cls.ERR_LOAD_CACHE_FAILED)
+        if success:
+            cls.update_cache()
 
     @classmethod
     def new_cache(cls, cache):
-        """New cache."""
+        """New settings."""
 
         default_cache = {'__format__': CACHE_FMT}
 
         try:
-            locked = False
-            with codecs.open(cache, "w", encoding="utf-8") as f:
-                assert portalocker.lock(f, portalocker.LOCK_EX), "Could not lock file."
-                locked = True
-                f.write(json.dumps(default_cache, sort_keys=True, indent=4, separators=(',', ': ')))
-                assert portalocker.unlock(f), "Could not unlock file."
-                locked = False
+            with cls.cache_lock.acquire(2):
+                with codecs.open(cache, "w", encoding="utf-8") as f:
+                    f.write(json.dumps(default_cache, sort_keys=True, indent=4, separators=(',', ': ')))
         except Exception:
-            try:
-                if locked:
-                    portalocker.unlock(f)
-            except Exception:
-                pass
+            # It's okay that we couldn't do it this time.
+            pass
 
         return default_cache
 
@@ -472,26 +447,15 @@ class Settings(object):
             settings = None
             cache = None
             if cls.settings_file is not None:
-                locked = False
                 try:
-                    with codecs.open(cls.settings_file, "r", encoding="utf-8") as f:
-                        assert portalocker.lock(f, portalocker.LOCK_SH), "Could not lock settings."
-                        locked = True
-                        settings = json.loads(f.read())
-                        assert portalocker.unlock(f), "could not unlock settings."
-                        locked = False
-                    with codecs.open(cls.cache_file, "r", encoding="utf-8") as f:
-                        assert portalocker.lock(f, portalocker.LOCK_SH), "Could not lock cache file."
-                        locked = True
-                        cache = json.loads(f.read())
-                        assert portalocker.unlock(f), "could not unlock cache file."
-                        locked = False
+                    with cls.settings_lock.acquire(2):
+                        with codecs.open(cls.settings_file, "r", encoding="utf-8") as f:
+                            settings = json.loads(f.read())
+                    with cls.cache_lock:
+                        with codecs.open(cls.cache_file, "r", encoding="utf-8") as f:
+                            cache = json.loads(f.read())
                 except Exception:
-                    try:
-                        if locked:
-                            portalocker.unlock(f)
-                    except Exception:
-                        pass
+                    pass
             if settings is not None:
                 cls.settings = settings
             if cache is not None:
@@ -1104,46 +1068,24 @@ class Settings(object):
         """Save settings."""
 
         try:
-            locked = False
             cls.settings['__format__'] = SETTINGS_FMT
-            with codecs.open(cls.settings_file, "w", encoding="utf-8") as f:
-                assert portalocker.lock(f, portalocker.LOCK_EX), "Could not lock settings."
-                locked = True
-                f.write(json.dumps(cls.settings, sort_keys=True, indent=4, separators=(',', ': ')))
-                assert portalocker.unlock(f), "could not unlock settings."
-                locked = False
+            with cls.settings_lock.acquire(2):
+                with codecs.open(cls.settings_file, "w", encoding="utf-8") as f:
+                    f.write(json.dumps(cls.settings, sort_keys=True, indent=4, separators=(',', ': ')))
         except Exception:
-            e = traceback.format_exc()
-            try:
-                if locked:
-                    portalocker.unlock(f)
-                errormsg(cls.ERR_SAVE_SETTINGS_FAILED)
-                error(e)
-            except Exception:
-                print(str(e))
+            errormsg(cls.ERR_SAVE_SETTINGS_FAILED)
 
     @classmethod
     def save_cache(cls):
         """Save cache."""
 
         try:
-            locked = False
             cls.cache['__format__'] = CACHE_FMT
-            with codecs.open(cls.cache_file, "w", encoding="utf-8") as f:
-                assert portalocker.lock(f, portalocker.LOCK_EX), "Could not lock cache file."
-                locked = True
-                f.write(json.dumps(cls.cache, sort_keys=True, indent=4, separators=(',', ': ')))
-                assert portalocker.unlock(f), "could not unlock cache file."
-                locked = False
+            with cls.cache_lock.acquire(2):
+                with codecs.open(cls.cache_file, "w", encoding="utf-8") as f:
+                    f.write(json.dumps(cls.cache, sort_keys=True, indent=4, separators=(',', ': ')))
         except Exception:
-            e = traceback.format_exc()
-            try:
-                if locked:
-                    portalocker.unlock(f)
-                errormsg(cls.ERR_SAVE_CACHE_FAILED)
-                error(e)
-            except Exception:
-                print(str(e))
+            errormsg(cls.ERR_SAVE_CACHE_FAILED)
 
     @classmethod
     def unload(cls):